#6_corrected_styles
