+++ components/StyledTextInput/styles.ts
@@ -1,0 +1,29 @@
+import { StyleSheet } from "react-native";
+
+export const styles = StyleSheet.create({
+    container: {
+      backgroundColor: '#785CEB',
+      borderRadius: 24,
+      padding: 16,
+      marginBottom: 16,
+      fontSize: 17,
+      minWidth:359,
+      minHeight:170,
+    },
+    containerWithFocus:{
+      backgroundColor: '#9379FF',
+      borderRadius: 24,
+      padding: 16,
+      marginBottom: 16,
+      fontSize: 17,
+      minWidth:359,
+      minHeight:170,
+    },
+    textInput: {
+  
+    },
+    textInputWithValue: {
+      color:'#FFFFFF'
+    },
+  });
+
