+++ ios/Podfile
@@ -58,4 +58,3 @@ target 'RN__test' do
     __apply_Xcode_12_5_M1_post_install_workaround(installer)
   end
 end
-pod 'BVLinearGradient', :path => '../node_modules/react-native-linear-gradient'
