+++ screens/styles.ts
@@ -1,0 +1,11 @@
+import { StyleSheet } from "react-native/types";
+
+export const styles = StyleSheet.create({
+    container: {
+      flex: 1,
+      backgroundColor: "#5A3ECD",
+      alignItems: "center",
+      justifyContent: "center",
+    },
+  });
+
